- Made fsizeof() return an off_t and changed the type of st_ucon64_t.file_size   to uint64_t (unsigned type large enough for the largest positive value that   off_t can have).  + Updated all references and the whole cascade of dependencies as long as they    made sense. - Changed type of st_ucon64_nfo_t.backup_header_len to unsigned int.  + Updated all client code where necessary. - Updated --id to be good enough to enable the help text. There are 2 collisions   between proper dumps in a GoodSNES 2.04 set: BS Super Earth Defense Force (J)   / Dokapon Gaiden - Honoo no Audition (J) and Lester the Unlikely (U) /   Sparkster (J). - file name -> filename. - Added --mgh. This does the same as --gd3, except that the resulting file will   have an MGD name and no header. - Added --smgh. Its main purpose is splitting HiROM games in MGH format.   Splitting is required to make them run with an MGH. - Fixed several buffer overruns. - Made documentation in mgd.h correspond with what the (verified)   implementation does (it does not attempt to create 9.3 filenames). - Extended mgd_make_name() to make adding a suffix optional.  + Updated all client code. - Added ugly UCON64_ISSET2() macro to compare an unsigned variable with   UCON64_UNKNOWN (-1). - Removed third parameter ("verbose") from build_cm_patterns() and replaced it   with state of misc.c, "int cm_verbose". cm_verbose is modified by -v and -q.   cm_verbose also controls verbosity of change_mem{2}().  + Updated client code of build_cm_patterns(). - st_ucon64_nfo_t.name is properly sanitized before being displayed. Corrupted   Genesis ROM images can no longer trigger "INTERNAL WARNING: vprintf2()   encountered an unsupported ANSI escape sequence" for the MinGW and Visual C++   ports. - Fixed bug in ucon64_pattern() where data was written to disk before it could   still be modified, resulting in changes to be omitted.

+++ trunk/ucon64/src/backup/gbx.c
@@ -1909,7 +1909,7 @@ gbx_read_sram (const char *filename, unsigned short parport, int start_bank)
 int
 gbx_write_sram (const char *filename, unsigned short parport, int start_bank)
 {
-  int bank, n_banks, n_bytes = 0, totalbytes, idx, i, j;
+  unsigned int bank, n_banks, n_bytes = 0, totalbytes, idx, i, j;
   time_t starttime;
   FILE *file;
 
@@ -1930,7 +1930,7 @@ gbx_write_sram (const char *filename, unsigned short parport, int start_bank)
     }
   else
     {
-      if (start_bank >= n_banks)
+      if ((unsigned int) start_bank >= n_banks)
         {
           fprintf (stderr, "ERROR: Bank must be a value 0 - %d (for this card)\n",
                    n_banks - 1);
